refactor: move main code files to src
