+++ legacy/autodetector.py
